+++ QuantumRefrigiz/DrawElefantQ.cs
@@ -29,7 +29,7 @@ namespace QuantumRefrigiz
         public bool OnlySelfT = false;
         public bool AStarGreedyHuristicT = false;
         public bool ArrangmentsChanged = false;
-        public static double MaxHuristicxE = -20000000000000000;
+        public static long MaxHuristicxE = -20000000000000000;
         public float Row, Column;
         public ThinkingQuantumChess[] ElefantThinkingQuantum = new ThinkingQuantumChess[AllDraw.ElefantMovments];
         public int[,] Table = null;
@@ -57,7 +57,7 @@ namespace QuantumRefrigiz
         public bool MaxFound(ref bool MaxNotFound)
         {
             
-                double a = ReturnHuristic();
+                int a = ReturnHuristic();
                 if (MaxHuristicxE < a)
                 {
                     Object O2 = new Object();
@@ -74,9 +74,9 @@ namespace QuantumRefrigiz
             MaxNotFound = true;
             return false;
         }
-        public double ReturnHuristic()
+        public int ReturnHuristic()
         {
-            double a = 0;
+            int a = 0;
             for (var ii = 0; ii < AllDraw.ElefantMovments; ii++)
                 
                     a += ElefantThinkingQuantum[ii].ReturnHuristic(-1, -1, Order, false);
@@ -234,9 +234,9 @@ namespace QuantumRefrigiz
                                 g.DrawImage(E[0], new Rectangle((int)(Row * (float)CellW), (int)(Column * (float)CellH), CellW, CellH));
                                 if (RingHalf)
                                 {
-                                    double Prob = 180;
+                                    int Prob = 180;
                                     if (AllDraw.Less != 0)
-                                        Prob = 180 * (AllDraw.Less / Double.MaxValue);
+                                        Prob = 180 * (AllDraw.Less / int.MaxValue);
                                     g.DrawArc(new Pen(new SolidBrush(Color.Red)), new Rectangle((int)((Row * (float)CellW)), (int)(Column * (float)CellH), CellW, CellH), -45, (int)Prob);
                                     if (LastRow != -1 && LastColumn != -1)
                                     {
@@ -269,9 +269,9 @@ namespace QuantumRefrigiz
                                 g.DrawImage(E[1], new Rectangle((int)(Row * (float)CellW), (int)(Column * (float)CellH), CellW, CellH));
                                 if (RingHalf)
                                 {
-                                    double Prob = 180;
+                                    int Prob = 180;
                                     if (AllDraw.Less != 0)
-                                        Prob = 180 * (AllDraw.Less / Double.MaxValue);
+                                        Prob = 180 * (AllDraw.Less / int.MaxValue);
                                     g.DrawArc(new Pen(new SolidBrush(Color.Red)), new Rectangle((int)((Row * (float)CellW)), (int)(Column * (float)CellH), CellW, CellH), -45, (int)Prob);
                                     if (LastRow != -1 && LastColumn != -1)
                                     {
