+++ QuantumRefrigiz/DrawKingQ.cs
@@ -31,7 +31,7 @@ namespace QuantumRefrigiz
         public bool AStarGreedyHuristicT = false;
 
         public bool ArrangmentsChanged = false;
-        public static double MaxHuristicxK = -20000000000000000;
+        public static long MaxHuristicxK = -20000000000000000;
         public float Row, Column;
         public Color color;
         public int[,] Table = null;
@@ -58,9 +58,9 @@ namespace QuantumRefrigiz
             K = null;
         }
 
-        public double ReturnHuristic()
+        public int ReturnHuristic()
         {
-            double a = 0;
+            int a = 0;
             for (var ii = 0; ii < AllDraw.KingMovments; ii++)
                 
                     a += KingThinkingQuantum[ii].ReturnHuristic(-1, -1, Order,false);
@@ -71,7 +71,7 @@ namespace QuantumRefrigiz
         public bool MaxFound(ref bool MaxNotFound)
         {
             
-                double a = ReturnHuristic();
+                int a = ReturnHuristic();
                 if (MaxHuristicxK < a)
                 {
                     Object O2 = new Object();
@@ -236,9 +236,9 @@ namespace QuantumRefrigiz
                                 g.DrawImage(K[0], new Rectangle((int)(Row * (float)CellW), (int)(Column * (float)CellH), CellW, CellH));
                                 if (RingHalf)
                                 {
-                                    double Prob = 180;
+                                    int Prob = 180;
                                     if (AllDraw.Less != 0)
-                                        Prob = 180 * (AllDraw.Less / Double.MaxValue);
+                                        Prob = 180 * (AllDraw.Less / int.MaxValue);
                                     g.DrawArc(new Pen(new SolidBrush(Color.Red)), new Rectangle((int)((Row * (float)CellW)), (int)(Column * (float)CellH), CellW, CellH), -45, (int)Prob);
                                     if (LastRow != -1 && LastColumn != -1)
                                     {
@@ -278,9 +278,9 @@ namespace QuantumRefrigiz
                                 g.DrawImage(K[1], new Rectangle((int)(Row * (float)CellW), (int)(Column * (float)CellH), CellW, CellH));
                                 if (RingHalf)
                                 {
-                                    double Prob = 180;
+                                    int Prob = 180;
                                     if (AllDraw.Less != 0)
-                                        Prob = 180 * (AllDraw.Less / Double.MaxValue);
+                                        Prob = 180 * (AllDraw.Less / int.MaxValue);
                                     g.DrawArc(new Pen(new SolidBrush(Color.Red)), new Rectangle((int)((Row * (float)CellW)), (int)(Column * (float)CellH), CellW, CellH), -45, (int)Prob);
                                     if (LastRow != -1 && LastColumn != -1)
                                     {
