+++ RfregitzD/DrawCastle.cs
@@ -28,7 +28,7 @@ namespace RefrigtzDLL
         public bool OnlySelfT = false;
         public bool AStarGreedyHuristicT = false;
         public bool ArrangmentsChanged = false;
-        public static double MaxHuristicxB = -20000000000000000;
+        public static long MaxHuristicxB = -20000000000000000;
         public float Row, Column;
         public Color color;
         public ThinkingChess[] CastleThinking = new ThinkingChess[AllDraw.CastleMovments];
@@ -56,7 +56,7 @@ namespace RefrigtzDLL
         public bool MaxFound(ref bool MaxNotFound)
         {
 
-            double a = ReturnHuristic();
+            int a = ReturnHuristic();
             if (MaxHuristicxB < a)
             {
                 MaxNotFound = false;
@@ -73,9 +73,9 @@ namespace RefrigtzDLL
             MaxNotFound = true;
             return false;
         }
-        public double ReturnHuristic()
+        public int ReturnHuristic()
         {
-            double a = 0;
+            int a = 0;
             for (var ii = 0; ii < AllDraw.CastleMovments; ii++)
 
                 a += CastleThinking[ii].ReturnHuristic(-1, -1, Order, false);
